removed unrequired concern and moved #silently_destroy_github_repository to a concern
