+++ app/models/concerns/assignment_repoable.rb
@@ -6,6 +6,8 @@ module AssignmentRepoable
   included do
     validates :github_repo_id, presence:   true
     validates :github_repo_id, uniqueness: true
+
+    before_destroy :silently_destroy_github_repository
   end
 
   def private?
@@ -20,5 +22,17 @@ module AssignmentRepoable
     return "No starter code provided" unless assignment.starter_code?
 
     github_repository.import_progress.status.humanize
+  end
+
+  # Internal: Attempt to destroy the GitHub repository.
+  #
+  # Returns true.
+  def silently_destroy_github_repository
+    return true if organization.blank?
+
+    organization.github_organization.delete_repository(github_repo_id)
+    true
+  rescue GitHub::Error
+    true
   end
 end
